+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog
 -------------------
 
 * Catch error when DB restore requested when DB service not available.
+* Fixed issue when retrieving docker compose configuration in versions of docker compose >= 2.
 
 3.13.1 (2021-08-06)
 -------------------