+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog
 
 * Catch error when DB restore requested when DB service not available.
 * Fixed issue when retrieving docker compose configuration in versions of docker compose >= 2.
+* Added sanity check to warn users if new and old configuration files are used at the same time. 
 
 3.13.1 (2021-08-06)
 -------------------