+++ CHANGELOG.rst
@@ -1,12 +1,13 @@
 Changelog
 =========
 
-3.14.0 (XXXX-XX-XX)
+3.14.0 (2021-10-06)
 -------------------
 
 * Catch error when DB restore requested when DB service not available.
 * Fixed issue when retrieving docker compose configuration in versions of docker compose >= 2.
-* Added sanity check to warn users if new and old configuration files are used at the same time. 
+* Added sanity check to warn users if new and old configuration files are used at the same time.
+* Updated requirements
 
 3.13.1 (2021-08-06)
 -------------------